Fix bug

+++ StreamProcess/src/main/java/applications/DataTypes/util/ISegmentIdentifier.java
@@ -3,7 +3,7 @@ package applications.DataTypes.util;
 public interface ISegmentIdentifier {
     Integer getXWay();
 
-    Short getSegment();
+    int getSegment();
 
     Short getDirection();
 }
